+++ src/__tests__/component.test.tsx
@@ -9,9 +9,7 @@ import {
   mockResourceGood,
   mockResourceBad,
 } from '../__fixtures__';
-import type { ResponsiveImageViewBag } from '../useResponsiveImageView';
-// eslint-disable-next-line import/no-named-as-default, import/no-named-as-default-member
-import ResponsiveImageView from '..';
+import { ResponsiveImageView, type ResponsiveImageViewBag } from '..';
 
 const expectedShape = expect.objectContaining<ResponsiveImageViewBag>({
   loading: expect.any(Boolean) as ResponsiveImageViewBag['loading'],
@@ -24,9 +22,9 @@ const expectedShape = expect.objectContaining<ResponsiveImageViewBag>({
 describe('rendering order (component > render > FAC > children > null)', () => {
   describe('renders component if provided', () => {
     it('function', () => {
-      const MyFunctionComponent = jest.fn(() => null);
-      const renderProp = jest.fn(() => null);
-      const children = jest.fn(() => null);
+      const MyFunctionComponent = jest.fn(() => <View />);
+      const renderProp = jest.fn(() => <View />);
+      const children = jest.fn(() => <View />);
       render(
         <ResponsiveImageView
           source={{ uri: mockUriGood }}
@@ -36,21 +34,21 @@ describe('rendering order (component > render > FAC > children > null)', () => {
           {children}
         </ResponsiveImageView>,
       );
-      expect(MyFunctionComponent).toHaveBeenCalledWith(expectedShape);
+      // the second argument is the context (undefined in this case), which we don't care about
+      expect(MyFunctionComponent).toHaveBeenCalledWith(expectedShape, undefined);
       expect(renderProp).not.toHaveBeenCalled();
       expect(children).not.toHaveBeenCalled();
     });
 
     it('class', () => {
-      const classRenderMethod = jest.fn((_props) => null);
+      const classRenderMethod = jest.fn((_props) => <View />);
       class MyClassComponent extends React.Component {
         render() {
           return classRenderMethod(this.props);
         }
       }
-
-      const renderProp = jest.fn(() => null);
-      const children = jest.fn(() => null);
+      const renderProp = jest.fn(() => <View />);
+      const children = jest.fn(() => <View />);
       render(
         <ResponsiveImageView
           source={{ uri: mockUriGood }}
@@ -67,8 +65,8 @@ describe('rendering order (component > render > FAC > children > null)', () => {
   });
 
   it('calls render if no component was provided', () => {
-    const renderProp = jest.fn(() => null);
-    const children = jest.fn(() => null);
+    const renderProp = jest.fn(() => <View />);
+    const children = jest.fn(() => <View />);
     render(
       <ResponsiveImageView source={{ uri: mockUriGood }} render={renderProp}>
         {children}
@@ -79,7 +77,7 @@ describe('rendering order (component > render > FAC > children > null)', () => {
   });
 
   it('calls children function if no component or render prop was provided', () => {
-    const children = jest.fn(() => null);
+    const children = jest.fn(() => <View />);
     render(<ResponsiveImageView source={{ uri: mockUriGood }}>{children}</ResponsiveImageView>);
     expect(children).toHaveBeenCalledWith(expectedShape);
   });
