+++ src/index.tsx
@@ -1,36 +1,216 @@
 import * as React from 'react';
-import { useResponsiveImageView, type ResponsiveImageViewProps } from './useResponsiveImageView';
+import {
+  Image,
+  type ViewProps,
+  type ImageProps,
+  type ImageURISource,
+  type ImageRequireSource,
+} from 'react-native';
+
+// Output
+export type ResponsiveImageViewBag = {
+  loading: boolean;
+  error: string;
+  retry: () => void;
+  getViewProps: (props?: ViewProps) => ViewProps;
+  getImageProps: (
+    props?: Omit<ImageProps, 'source'> & { source?: ImageProps['source'] },
+  ) => ImageProps;
+};
+
+// Hook options
+export type UseResponsiveImageViewOptions = {
+  source: ImageURISource | ImageRequireSource;
+  aspectRatio?: number;
+  onLoad?: () => void;
+  onError?: (errMessage: string) => void;
+};
+
+// Component props
+export type ResponsiveImageViewProps = UseResponsiveImageViewOptions & {
+  component?: React.ComponentType<any>;
+  render?: (bag: ResponsiveImageViewBag) => React.JSX.Element;
+  children?: ((bag: ResponsiveImageViewBag) => React.JSX.Element) | React.ReactNode;
+};
+
+// A cancelable version of Image.getSize, adapted from
+// https://github.com/kodefox/react-native-flex-image
+function getImageSize(
+  uri: Parameters<typeof Image.getSize>[0],
+  onImageSizeSuccess: Parameters<typeof Image.getSize>[1],
+  onImageSizeFailure: Parameters<typeof Image.getSize>[2],
+) {
+  let totallyCanceled = false;
+
+  Image.getSize(
+    uri,
+    (width, height) => {
+      if (!totallyCanceled) {
+        onImageSizeSuccess(width, height);
+      }
+    },
+    (err) => {
+      if (!totallyCanceled) {
+        onImageSizeFailure?.(err);
+      }
+    },
+  );
+
+  return {
+    cancel: () => {
+      totallyCanceled = true;
+    },
+  };
+}
+
+type State = {
+  loading: boolean;
+  error: string;
+  retryCount: number;
+  aspectRatio: number | undefined;
+};
+
+type Action =
+  | { type: 'SUCCESS'; payload: number }
+  | { type: 'FAILURE'; payload: string }
+  | { type: 'RETRY' };
+
+const initialState: State = { loading: true, error: '', retryCount: 0, aspectRatio: undefined };
+
+function reducer(state: State, action: Action) {
+  switch (action.type) {
+    case 'SUCCESS': {
+      return {
+        ...initialState,
+        loading: false,
+        retryCount: state.retryCount,
+        aspectRatio: action.payload,
+      };
+    }
+    case 'FAILURE': {
+      return {
+        ...initialState,
+        loading: false,
+        error: action.payload,
+        retryCount: state.retryCount,
+      };
+    }
+    case 'RETRY': {
+      return { ...initialState, retryCount: state.retryCount + 1 };
+    }
+    /* istanbul ignore next: this will never happen */
+    default: {
+      throw new Error('Unexpected action type');
+    }
+  }
+}
 
 function defaultOnLoad() {}
 function defaultOnError(_: string) {}
 
-function isFunctionComponent(
-  component: React.ComponentType | React.FunctionComponent,
-): component is React.FunctionComponent {
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-  return !component.prototype?.isReactComponent;
+export function useResponsiveImageView({
+  aspectRatio: controlledAspectRatio,
+  source: initialSource,
+  onLoad = defaultOnLoad,
+  onError = defaultOnError,
+}: UseResponsiveImageViewOptions): ResponsiveImageViewBag {
+  if (!initialSource) {
+    throw new Error('"source" is required');
+  }
+
+  const [state, dispatch] = React.useReducer(reducer, initialState);
+
+  function retry() {
+    dispatch({ type: 'RETRY' });
+  }
+
+  function isAspectRatioControlled() {
+    return controlledAspectRatio !== undefined;
+  }
+
+  function getAspectRatio() {
+    return isAspectRatioControlled() ? controlledAspectRatio : state.aspectRatio;
+  }
+
+  function getImageProps({
+    source,
+    style = {},
+    ...props
+  }: Parameters<ResponsiveImageViewBag['getImageProps']>[0] = {}) {
+    const imageProps: ImageProps = {
+      source: initialSource,
+      style: [style, { height: '100%', width: '100%' }],
+      ...props,
+    };
+    return imageProps;
+  }
+
+  function getViewProps({
+    style = {},
+    ...props
+  }: Parameters<ResponsiveImageViewBag['getViewProps']>[0] = {}) {
+    return { style: [style, { aspectRatio: getAspectRatio() }], ...props };
+  }
+
+  React.useEffect(() => {
+    let pendingGetImageSize = { cancel: /* istanbul ignore next: just a stub  */ () => {} };
+
+    function handleImageSizeSuccess(width: number, height: number) {
+      onLoad();
+      dispatch({ type: 'SUCCESS', payload: width / height });
+    }
+
+    function handleImageSizeFailure(err: any) {
+      const errMessage =
+        err instanceof Error ? err.message : /* istanbul ignore next */ String(err);
+      onError(errMessage);
+      dispatch({ type: 'FAILURE', payload: errMessage });
+    }
+
+    if (typeof initialSource === 'object' && initialSource.uri) {
+      // Retrieve image dimensions from URI
+      pendingGetImageSize = getImageSize(
+        initialSource.uri,
+        handleImageSizeSuccess,
+        handleImageSizeFailure,
+      );
+    } else {
+      // Retrieve image dimensions from imported resource
+      const imageSource = Image.resolveAssetSource(initialSource);
+      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+      if (imageSource) {
+        handleImageSizeSuccess(imageSource.width, imageSource.height);
+      } else {
+        handleImageSizeFailure(new Error('Failed to retrieve image dimensions.'));
+      }
+    }
+
+    return () => {
+      pendingGetImageSize.cancel();
+    };
+    // Using JSON.stringify here because the `source` parameter can be a nested
+    // object. The alternative is requiring the user to memoize it, by why make
+    // them do that when we don't have to? (Note: they already have to memoize
+    // onLoad and onError, but those are much less likely to be used.)
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [JSON.stringify(initialSource), onLoad, onError, state.retryCount]);
+
+  return { loading: state.loading, error: state.error, retry, getViewProps, getImageProps };
 }
 
-function ResponsiveImageView({
+export function ResponsiveImageView({
+  source,
   component: Component = undefined,
   render = undefined,
   children = undefined,
   aspectRatio = undefined,
-  source,
   onLoad = defaultOnLoad,
   onError = defaultOnError,
-}: ResponsiveImageViewProps): React.ReactElement<ResponsiveImageViewProps> | null {
+}: ResponsiveImageViewProps): React.JSX.Element | null {
   const bag = useResponsiveImageView({ aspectRatio, source, onLoad, onError });
 
   // component injection
   if (Component) {
-    // function component
-    if (isFunctionComponent(Component)) {
-      // casting to synchronous as RN doesn't support async function components (RSC)
-      return Component(bag) as React.ReactElement<ResponsiveImageViewProps>;
-    }
-
-    // class component
     return <Component {...bag} />;
   }
 
@@ -53,6 +233,3 @@ function ResponsiveImageView({
 }
 
 ResponsiveImageView.displayName = 'ResponsiveImageView';
-
-export { useResponsiveImageView };
-export default ResponsiveImageView;
